separation of the MoviesList component into a MoviesContainer and a MoviesList
